split eval selector from button
